#1106 - Switch from jmockit to mockito in webanno-brat
- Improve testability of the PreRenderer
- Switch tests to use mockito
