+++ webanno-brat/src/test/java/de/tudarmstadt/ukp/clarin/webanno/brat/message/GetCollectionInformationResponseTest.java
@@ -1,0 +1,100 @@
+/*
+ * Copyright 2018
+ * Ubiquitous Knowledge Processing (UKP) Lab and FG Language Technology
+ * Technische Universität Darmstadt
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.tudarmstadt.ukp.clarin.webanno.brat.message;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.linesOf;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.junit.Test;
+
+import de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst;
+import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
+import de.tudarmstadt.ukp.clarin.webanno.model.Tag;
+import de.tudarmstadt.ukp.clarin.webanno.model.TagSet;
+import de.tudarmstadt.ukp.clarin.webanno.support.JSONUtil;
+
+public class GetCollectionInformationResponseTest
+{
+    /**
+     * generate BRAT JSON for the collection informations
+     *
+     * @throws IOException
+     *             if an I/O error occurs.
+     */
+    @Test
+    public void testGenerateBratJsonGetCollection() throws IOException
+    {
+        String jsonFilePath = "target/test-output/output_cas_to_json_collection.json";
+
+        GetCollectionInformationResponse collectionInformation = 
+                new GetCollectionInformationResponse();
+
+        List<AnnotationLayer> layerList = new ArrayList<>();
+
+        AnnotationLayer layer = new AnnotationLayer();
+        layer.setId(1l);
+        layer.setDescription("span annoattion");
+        layer.setName("pos");
+        layer.setType(WebAnnoConst.SPAN_TYPE);
+
+        TagSet tagset = new TagSet();
+        tagset.setId(1l);
+        tagset.setDescription("pos");
+        tagset.setLanguage("de");
+        tagset.setName("STTS");
+
+        Tag tag = new Tag();
+        tag.setId(1l);
+        tag.setDescription("noun");
+        tag.setName("NN");
+        tag.setTagSet(tagset);
+
+        layerList.add(layer);
+
+        collectionInformation.addCollection("/Collection1/");
+        collectionInformation.addCollection("/Collection2/");
+        collectionInformation.addCollection("/Collection3/");
+
+        collectionInformation.addDocument("/Collection1/doc1");
+        collectionInformation.addDocument("/Collection2/doc1");
+        collectionInformation.addDocument("/Collection3/doc1");
+        collectionInformation.addDocument("/Collection1/doc2");
+        collectionInformation.addDocument("/Collection2/doc2");
+        collectionInformation.addDocument("/Collection3/doc2");
+
+        collectionInformation.setSearchConfig(new ArrayList<>());
+
+        List<String> tagSetNames = new ArrayList<>();
+        tagSetNames.add(de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.POS);
+        tagSetNames.add(de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.DEPENDENCY);
+        tagSetNames.add(de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.NAMEDENTITY);
+        tagSetNames.add(de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.COREFERENCE);
+        tagSetNames.add(de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.COREFRELTYPE);
+
+        JSONUtil.generatePrettyJson(collectionInformation, new File(jsonFilePath));
+
+        assertThat(
+                linesOf(new File("src/test/resources/output_cas_to_json_collection_expected.json"),
+                        "UTF-8")).isEqualTo(linesOf(new File(jsonFilePath), "UTF-8"));
+    }
+}
